Se agregaron varias clases de dominio, servicios y controladores, pero el proyecto quedara en stand by de momento

+++ src/main/java/com/arg/ecommerce/dao/ProductDao.java
@@ -3,6 +3,6 @@ package com.arg.ecommerce.dao;
 import com.arg.ecommerce.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductDao extends JpaRepository<Product, Integer> {
+public interface ProductDao extends JpaRepository<Product, Long> {
 
 }
