+++ src/main/java/com/arg/ecommerce/dao/ShoppingSessionDao.java
@@ -3,6 +3,6 @@ package com.arg.ecommerce.dao;
 import com.arg.ecommerce.domain.ShoppingSession;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShoppingSessionDao extends JpaRepository<ShoppingSession, Integer> {
+public interface ShoppingSessionDao extends JpaRepository<ShoppingSession, Long> {
 
 }
